pyproject.toml
[project]
name = "determinagent"
version = "0.10.0"
description = "CLI-First deterministic multi-agent orchestration library"
authors = [{ name = "DeterminAgent Team", email = "victor@experto.ai" }]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["ai", "agents", "cli", "orchestration", "langgraph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langgraph>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "markdown>=3.0",
    "beautifulsoup4>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "types-pyyaml>=6.0",
    "types-markdown>=3.0",
    "types-beautifulsoup4>=4.9.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings-python>=1.0.0",
    "mkdocs-htmlproofer-plugin>=1.0.0",
    # Security scanning
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
]

# Global CLI removed in favor of template flows
#[project.scripts]
#determinagent = "determinagent.cli:cli_main"

[project.urls]
Homepage = "https://github.com/determinagent/determinagent"
Documentation = "https://github.com/determinagent/determinagent#readme"
Repository = "https://github.com/determinagent/determinagent"
Issues = "https://github.com/determinagent/determinagent/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
    "/determinagent",
]

[tool.hatch.build.targets.wheel]
packages = ["determinagent"]

# ============================================================================
# Ruff Configuration
# ============================================================================
[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "__pycache__",
    "*.pyc",
    "legacy",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Function call in argument defaults
    "C901",   # Too complex
    "ARG001", # Unused function argument (often needed for ABCs)
    "ARG002", # Unused method argument
]

[tool.ruff.lint.isort]
known-first-party = ["determinagent"]

# ============================================================================
# Black Configuration
# ============================================================================
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
exclude = '''
/(
    \.git
    | \.mypy_cache
    | __pycache__
    | legacy
)/
'''

# ============================================================================
# Mypy Configuration
# ============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true
exclude = [
    "legacy/",
    "tests/",
]

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "bs4.*",
]
ignore_missing_imports = true

# ============================================================================
# Pytest Configuration
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (require CLI tools)",
    "smoke: Smoke tests (verify basic imports and flow structure)",
    "slow: Slow tests",
]

# ============================================================================
# Coverage Configuration
# ============================================================================
[tool.coverage.run]
source = ["determinagent"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true

# ============================================================================
# Bandit Security Scanner Configuration
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", "legacy", ".venv", "venv"]
skips = [
    "B101",  # assert_used - acceptable in library code for invariants
    "B404",  # import_subprocess - required for CLI orchestration
    "B603",  # subprocess_without_shell_equals_true - intentional design
    "B607",  # start_process_with_partial_path - CLI tools are user-configured
]
targets = ["determinagent"]

